% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_tf.R
\name{briere}
\alias{briere}
\title{briere}
\usage{
briere(temp_values = c(25), T0 = 2.248952, Tm = 40.13383, c = 0.000271964)
}
\arguments{
\item{temp_values}{TBA}

\item{T0}{TBA}

\item{Tm}{TBA}

\item{c}{TBA
'}
}
\description{
Briere function for temperature transformation
}
\details{
TBA
}
