% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{case_data_calculate}
\alias{case_data_calculate}
\title{case_data_calculate}
\usage{
case_data_calculate(
  model_data = list(),
  n_p = 1,
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe = 1,
  p_rep_death = 1,
  output_type = "annual",
  deterministic = FALSE
)
}
\arguments{
\item{model_data}{SEIRV output of Model_Run and similar functions}

\item{n_p}{Particle to select from model_data}

\item{p_severe_inf}{Probability of an infection being severe}

\item{p_death_severe_inf}{Probability of a severe infection resulting in death}

\item{p_rep_severe}{Probability of reporting of a severe but non-fatal infection}

\item{p_rep_death}{Probability of reporting of a fatal infection}

\item{output_type}{Type of output to produce:
"annual" - Total reported cases and reported deaths by year
"pts" - Total reported cases and reported deaths for every time point, summed over age groups
"full" - Reported cases and reported deaths for every time point and age group}

\item{deterministic}{Indicates whether to calculate results deterministically (TRUE) or stochastically (FALSE)
'}
}
\description{
Calculate reported case data from SEIRV model output
}
\details{
[TBA]
}
