% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{case_data_compare}
\alias{case_data_compare}
\title{case_data_compare}
\usage{
case_data_compare(model_values, obs_values)
}
\arguments{
\item{model_values}{Modelled reported case or death values}

\item{obs_values}{Observed or template case or death values
'}
}
\description{
Compare modelled and observed case or death data using negative binomial function
}
\details{
Compares modelled data (from dataset generation functions) on reported cases or deaths
per year with observed data, calculating logarithmiclikelihood of observing the latter given the former,
using a negative binomial formula.
}
