% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_param_labels}
\alias{create_param_labels}
\title{create_param_labels}
\usage{
create_param_labels(enviro_data = NULL, extra_estimated_params = c("vacc_eff"))
}
\arguments{
\item{enviro_data}{Environmental data frame, containing only relevant environmental covariate values for regions of interest}

\item{extra_estimated_params}{Vector of strings listing variable parameters besides ones determining FOI/R0 (may include
vaccine efficacy and/or infection/death reporting probabilities and/or Brazil FOI adjustment factor)}
}
\description{
Apply names to the parameters in a set used for data matching and parameter fitting
}
\details{
Takes in environmental covariate data along with names of additional parameters (vaccine efficacy
and reporting probabilities) and generates list of names for parameter set to use as input for fitting functions
}
