% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{enviro_data_var_check}
\alias{enviro_data_var_check}
\title{enviro_data_var_check}
\usage{
enviro_data_var_check(enviro_data_var)
}
\arguments{
\item{enviro_data_var}{TBA}
}
\description{
Check format of variable environmental data
}
\details{
TBA
}
