% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{enviro_data_var_truncate}
\alias{enviro_data_var_truncate}
\title{enviro_data_var_truncate}
\usage{
enviro_data_var_truncate(
  enviro_data_var = list(),
  regions_new = c(),
  t_pts_new = NULL
)
}
\arguments{
\item{enviro_data_var}{TBA}

\item{regions_new}{Vector of regions (subset of enviro_data_var$regions)}

\item{t_pts_new}{TBA}
}
\description{
Truncate variable environmental data for shorter set of regions and/or time points
}
\details{
TBA
}
