% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{epi_param_calc}
\alias{epi_param_calc}
\title{epi_param_calc}
\usage{
epi_param_calc(
  coeffs_const = c(),
  coeffs_var = c(),
  enviro_data_const = data.frame(),
  enviro_data_var = NULL
)
}
\arguments{
\item{coeffs_const}{Vector of coefficients of time-invariant covariates}

\item{coeffs_var}{Vector of coefficients of time-varying covariates}

\item{enviro_data_const}{Data frame of time-invariant environmental covariate values, with region labels in first
column and one row per region}

\item{enviro_data_var}{List containing data on time-varying environmental covariates: [TBA]
'}
}
\description{
Calculate FOI_spillover or R0 values from environmental covariates and coefficients
}
\details{
Takes in environmental covariate values for one or more regions and coefficients of environmental covariates
 and calculates epidemiological parameter values via matrix multiplication. Environmental covariates may be constant
 or time-varying; constant and time-varying sets of values are supplied as separate input variables, as are coefficients
 of constant and time-varying covariates.

 The function can accept [TBA]
}
