% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_gen.R
\name{Generate_Dataset}
\alias{Generate_Dataset}
\title{Generate_Dataset}
\usage{
Generate_Dataset(
  FOI_values = c(),
  R0_values = c(),
  input_data = list(),
  template = list(sero = NULL, case = NULL, xref_sero = NULL, xref_case = NULL),
  vaccine_efficacy = 1,
  time_inc = 1,
  mode_start = 1,
  mode_time = 0,
  n_reps = 1,
  deterministic = FALSE,
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe = 1,
  p_rep_death = 1,
  use_node_cluster = FALSE,
  n_nodes = NULL,
  output_frame = FALSE,
  seed = NULL,
  region_grouping = NULL,
  mode_grouping = 1
)
}
\arguments{
\item{FOI_values}{Array of values of force of infection due to spillover from sylvatic reservoir by region + time}

\item{R0_values}{Array of values of basic reproduction number for human-human transmission by region + time}

\item{input_data}{List of population and vaccination data for multiple regions in standard format [TBA]}

\item{template}{TBA \cr
sero: Seroprevalence data template - data frame with region, year, minimum/maximum age, vc_factor [TBA]
and no. samples \cr
case: Annual reported case/death data template - data frame with region and year \cr}

\item{vaccine_efficacy}{Fractional vaccine efficacy}

\item{time_inc}{Time increment in days to use in model (should be either 1.0, 2.5 or 5.0 days)}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination \cr
If mode_start = 0, only vaccinated individuals \cr
If mode_start = 1, shift some non-vaccinated individuals into R to give herd immunity (stratified by age) \cr
If mode_start = 2, use SEIRV input in list from previous run(s)}

\item{mode_time}{Type of time dependence of FOI_spillover and R0 to be used: \cr
If mode_time = 0, no time variation (constant values)\cr
If mode_time = 1, FOI/R0 vary annually without seasonality (no. values = no. years to consider) \cr
If mode_time = 2, FOI/R0 vary with monthly seasonality without inter-annual variation (no. values = 12) \cr
If mode_time = 3, FOI/R0 vary with daily seasonality without inter-annual variation (no. values = 365/dt) \cr
If mode_time = 4, FOI/R0 vary annually with monthly seasonality (no. values = 12*no. years to consider) \cr
If mode_time = 5, FOI/R0 vary annually with daily seasonality (no. values = (365/dt)*no. years to consider)}

\item{n_reps}{no. stochastic repetitions}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE}

\item{p_severe_inf}{Probability of an infection being severe}

\item{p_death_severe_inf}{Probability of a severe infection resulting in death}

\item{p_rep_severe}{Probability of reporting of a severe but non-fatal infection}

\item{p_rep_death}{Probability of reporting of a fatal infection}

\item{use_node_cluster}{TRUE/FALSE - set model to run in parallel using cluster if TRUE}

\item{n_nodes}{no. nodes to use in cluster (if use_node_cluster = TRUE)}

\item{output_frame}{TRUE/FALSE - indicate whether to output a complete data frame of results in template format
(if TRUE) or calculated values only (if FALSE)}

\item{seed}{Optional random seed value; set to NULL to omit.}

\item{region_grouping}{TBA}

\item{mode_grouping}{TBA
'}
}
\description{
Generate dataset [TBA]
}
\details{
This function is used to [TBA]

[TBA - Explanation of breakdown of regions to model and how to set lengths of FOI_values and R0_values]
}
