% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_generate.R
\name{Generate_Multiple_Datasets}
\alias{Generate_Multiple_Datasets}
\title{Generate_Multiple Datasets}
\usage{
Generate_Multiple_Datasets(
  input_data = list(),
  FOI_values = list(),
  R0_values = list(),
  sero_template = NULL,
  case_template = NULL,
  vaccine_efficacy = c(1),
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe = c(1),
  p_rep_death = c(1),
  mode_start = 1,
  start_SEIRV = NULL,
  dt = 1,
  n_reps = 1,
  deterministic = FALSE,
  mode_parallel = "none",
  cluster = NULL
)
}
\arguments{
\item{input_data}{List of population and vaccination data for multiple regions in standard format [TBA]}

\item{FOI_values}{Array of sets of values of the force of infection due to spillover from sylvatic reservoir;
dimensions [TBA]}

\item{R0_values}{Array of values of the basic reproduction number for human-human transmission;
dimensions [TBA]}

\item{sero_template}{Seroprevalence data template - data frame with region, year, minimum/maximum age, vc_factor [TBA]
and number of samples}

\item{case_template}{Annual reported case/death data template - data frame with region and year}

\item{vaccine_efficacy}{Vector of values of vaccine efficacy [TBA]}

\item{p_severe_inf}{Probability of an infection being severe}

\item{p_death_severe_inf}{Probability of a severe infection resulting in death}

\item{p_rep_severe}{Vector of values of probability of reporting of a severe but non-fatal infection [TBA]}

\item{p_rep_death}{Vetor of value of probability of reporting of a fatal infection [TBA]}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input (list of datasets, one per region)}

\item{dt}{Time increment in days to use in model (should be either 1.0, 2.5 or 5.0 days)}

\item{n_reps}{number of stochastic repetitions}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE}

\item{mode_parallel}{Set mode for parallelization, if any:
If mode_parallel="none", no parallelization
If mode_parallel="pars_multi", all regions run in parallel for same time period with same output type
If mode_parallel="clusterMap", all regions run in parallel with different time periods and output types}

\item{cluster}{Cluster of threads to use if mode_parallel="clusterMap"
'}
}
\description{
Generate multiple datasets for multiple sets of input parameter values
}
\details{
This function is used to generate annual serological and/or case/death data based on templates for
  multiple sets of parameters (spillover FOI, R0, case/death reporting probabilities, vaccine efficacy)
}
