% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{input_data_truncate}
\alias{input_data_truncate}
\title{input_data_truncate}
\usage{
input_data_truncate(input_data = list(), regions_new = NULL, years_new = NULL)
}
\arguments{
\item{input_data}{List of population and vaccination data for multiple regions (created using create_input_data()
function and usually loaded from an RDS file)}

\item{regions_new}{Vector of regions (subset of input_data$region_labels)}

\item{years_new}{Vector of years (subset of input_data$years_labels)}
}
\description{
Truncate input data list for shorter set of regions and/or years
}
\details{
amends a list of population and vaccination data used as input for other functions,
  cross-referencing it with seroprevalence and/or case data and adding connection information.

  function and usually loaded from an RDS file)
  positives
  deaths

TBA
}
