% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{m_prelim_fit}
\alias{m_prelim_fit}
\title{m_prelim_fit}
\usage{
m_prelim_fit(
  fit_data = list(),
  packer = NULL,
  prior = NULL,
  FOI_R0_prior_data = list(),
  pars_var = list(),
  env_covar_values = list(),
  n_steps = 1,
  n_iterations = 10,
  n_bounds = 10,
  n_particles = 1,
  n_threads = 1,
  seed = NULL
)
}
\arguments{
\item{fit_data}{Fitting data created using fit_data_setup()}

\item{packer}{Packer created using packer_setup()}

\item{prior}{Prior function created using prior_setup()}

\item{FOI_R0_prior_data}{Data frame of prior data for FOI/R0}

\item{pars_var}{Data frame of information on varied parameters, created using pars_var_setup()}

\item{env_covar_values}{Environmental covariate values (TBC)}

\item{n_steps}{Number of times to run estimation cycle}

\item{n_iterations}{Number of iterations to run per cycle}

\item{n_bounds}{Number of iterations (ones giving highest posterior likelihood) to use to establish bounds
for next cycle}

\item{n_particles}{Number of particles}

\item{n_threads}{Number of threads}

\item{seed}{Random seed (set to NULL if unused)}
}
\description{
Function for preliminary estimation of parameter values
}
\details{
TBA
}
