% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{m_sample}
\alias{m_sample}
\title{m_sample}
\usage{
m_sample(
  fit_data = list(),
  packer = NULL,
  prior = NULL,
  FOI_R0_prior_data = list(),
  pars_var = list(),
  initial = list(),
  v = c(),
  n_chains = 1,
  n_iterations = 10,
  n_particles = 1,
  n_threads = 1,
  seed = NULL,
  parallel = FALSE,
  output_file = ""
)
}
\arguments{
\item{fit_data}{TBA}

\item{packer}{TBA}

\item{prior}{TBA}

\item{FOI_R0_prior_data}{TBA}

\item{pars_var}{TBA}

\item{initial}{TBA}

\item{v}{TBA}

\item{n_chains}{TBA}

\item{n_iterations}{TBA}

\item{n_particles}{TBA}

\item{n_threads}{TBA}

\item{seed}{TBA}

\item{parallel}{TBA}

\item{output_file}{TBA}
}
\description{
TBA
}
\details{
TBA
}
