% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_params_data_create}
\alias{mcmc_params_data_create}
\title{mcmc_params_data_create}
\usage{
mcmc_params_data_create(
  covar_names = c("Var1", "Var2"),
  add_est_param_names = "vaccine_efficacy"
)
}
\arguments{
\item{covar_names}{TBA}

\item{add_est_param_names}{TBA}
}
\description{
Create data frame of parameter data for input into MCMC() and mcmc_prelim_fit()
}
\details{
TBA
}
