% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_prelim_fit}
\alias{mcmc_prelim_fit}
\title{mcmc_prelim_fit}
\usage{
mcmc_prelim_fit(
  n_iterations = 1,
  n_param_sets = 1,
  n_bounds = 1,
  params_data = list(),
  input_data = list(),
  obs_sero_data = list(),
  obs_case_data = list(),
  mode_grouping = 2,
  mode_start = 1,
  time_inc = 1,
  n_reps = 1,
  enviro_data_const = list(),
  enviro_data_var = list(),
  deterministic = TRUE,
  mode_time = 0,
  mode_parallel = FALSE,
  cluster = NULL,
  plot_graphs = FALSE,
  seed = NULL
)
}
\arguments{
\item{n_iterations}{= Number of times to run and adjust maximum/minimum}

\item{n_param_sets}{= Number of parameter sets to run in each iteration}

\item{n_bounds}{= Number of parameter sets (with highest likelihood values) to take at each iteration to create new
maximum/minimum values}

\item{params_data}{TBA}

\item{input_data}{List of population and vaccination data for multiple regions (created using data input creation
code and usually loaded from RDS file)}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no. cases/no.
deaths}

\item{mode_grouping}{TBA}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination \cr
If mode_start = 0, only vaccinated individuals \cr
If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity (stratified by age) \cr}

\item{time_inc}{time increment in days (must be 1 or 5)}

\item{n_reps}{Number of repetitions}

\item{enviro_data_const}{Data frame of values of constant environmental covariates (columns) by region (rows)}

\item{enviro_data_var}{List containing values of time-varying environmental covariates (TBA)}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE}

\item{mode_time}{TBA}

\item{mode_parallel}{TRUE/FALSE - indicate whether to use parallel processing on supplied cluster for speed}

\item{cluster}{Cluster of threads to use if mode_parallel = TRUE}

\item{plot_graphs}{TRUE/FALSE - plot graphs of evolving parameter space}

\item{seed}{TBA
'}
}
\description{
Test multiple sets of parameters randomly drawn from range between maximum and minimum
values in order to find approximate values giving maximum posterior likelihood
}
\details{
This function is used to estimate the model parameter values giving maximum posterior likelihood; it is
primarily intended to be used to generate initial parameter values for Markov Chain Monte Carlo fitting (using
the mcmc() function).
}
