% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc2.R
\name{mcmc_prelim_fit2}
\alias{mcmc_prelim_fit2}
\title{mcmc_prelim_fit2}
\usage{
mcmc_prelim_fit2(
  n_iterations = 1,
  n_param_sets = 1,
  n_bounds = 1,
  type = NULL,
  log_params_min = NULL,
  log_params_max = NULL,
  input_data = list(),
  obs_sero_data = list(),
  obs_case_data = list(),
  mode_start = 0,
  prior_settings = list(type = "zero"),
  dt = 1,
  n_reps = 1,
  enviro_data = NULL,
  R0_fixed_values = c(),
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  add_values = list(vaccine_efficacy = 1, p_rep_severe_af = 1, p_rep_death_af = 1,
    p_rep_severe_sa = 1, p_rep_death_sa = 1, m_FOI_Brazil = 1),
  deterministic = TRUE,
  mode_parallel = "none",
  cluster = NULL
)
}
\arguments{
\item{n_iterations}{TBA}

\item{n_param_sets}{TBA}

\item{n_bounds}{TBA}

\item{type}{TBA}

\item{log_params_min}{TBA}

\item{log_params_max}{TBA}

\item{input_data}{TBA}

\item{obs_sero_data}{TBA}

\item{obs_case_data}{TBA}

\item{mode_start}{TBA}

\item{prior_settings}{TBA}

\item{dt}{TBA}

\item{n_reps}{TBA}

\item{enviro_data}{TBA}

\item{R0_fixed_values}{TBA}

\item{p_severe_inf}{TBA}

\item{p_death_severe_inf}{TBA}

\item{add_values}{TBA}

\item{deterministic}{TBA}

\item{mode_parallel}{TBA}

\item{cluster}{TBA
'}
}
\description{
TBA
}
\details{
TBA
}
