% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Model_Run}
\alias{Model_Run}
\title{Model_Run}
\usage{
Model_Run(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  year0 = 1940,
  years_data = c(1940:1941),
  mode_start = 0,
  vaccine_efficacy = 1,
  start_SEIRV = list(),
  dt = 1,
  output_type = "full"
)
}
\arguments{
\item{FOI_spillover}{Force of infection due to spillover from sylvatic reservoir}

\item{R0}{Basic reproduction number for urban spread of infection}

\item{vacc_data}{Vaccination coverage in each age group by year}

\item{pop_data}{Population in each age group by year}

\item{year0}{First year in population/vaccination data}

\item{years_data}{Incremental vector of years denoting start of data saving through to year at which to stop, i.e. if
data from 1945 to 1950 is required, years_data should be c(1945:1951)}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRVC input in list from previous run(s)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{dt}{Time increment in days to use in model (should be either 1.0 or 5.0 days)}

\item{output_type}{Output data to save:
 "full" for all SEIRV data plus day, year, total FOI
 "case" for just year and number of new infections
'}
}
\description{
Run SEIRV model
}
\details{
Accepts epidemiological + population parameters and model settings; runs SEIRV model
for one region over a specified time period for a number of particles/threads and outputs time-dependent SEIRV
values, infection numbers and total force of infection values.
}
