% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_OD.R
\name{Model_Run_Multi_Region}
\alias{Model_Run_Multi_Region}
\title{Model_Run_Multi_Region}
\usage{
Model_Run_Multi_Region(
  FOI_spillover = c(),
  R0 = c(),
  vacc_data = list(),
  pop_data = list(),
  years_data = c(1940:1941),
  start_SEIRV = list(),
  output_type = "full",
  year0 = 1940,
  mode_start = 0,
  vaccine_efficacy = 1,
  dt = 1,
  n_particles = 1,
  n_threads = 1,
  deterministic = FALSE
)
}
\arguments{
\item{FOI_spillover}{Vector of values of force of infection due to spillover from sylvatic reservoir}

\item{R0}{Vector of values of basic reproduction number for urban spread of infection}

\item{vacc_data}{Vaccination coverage in each region and age group by year}

\item{pop_data}{Population in each region and age group by year}

\item{years_data}{Incremental vector of years denoting years for which to save data}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{output_type}{Type of data to output: "full" = SEIRVC, "case" = C only, "sero" = SEIRV only}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{dt}{Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)}

\item{n_particles}{number of particles to use}

\item{n_threads}{number of threads to use}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE
'}
}
\description{
Run SEIRV model for multiple regions
}
\details{
Accepts epidemiological + population parameters and model settings; runs SEIRV model
for one region over a specified time period for a number of particles/threads and outputs time-dependent SEIRV
values, infection numbers and total force of infection values.
}
