% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{Model_Run_Threaded}
\alias{Model_Run_Threaded}
\title{Model_Run_Threaded}
\usage{
Model_Run_Threaded(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  years_data = c(),
  flag_case = 0,
  flag_sero = 0,
  year0 = 1940,
  mode_start = 0,
  vaccine_efficacy = 1,
  dt = 1
)
}
\arguments{
\item{FOI_spillover}{Force of infection due to spillover from sylvatic reservoir}

\item{R0}{Basic reproduction number for urban spread of infection}

\item{vacc_data}{Vaccination coverage in each age group by year}

\item{pop_data}{Population in each age group by year}

\item{years_data}{Incremental vector of years for which to output SEIRV data}

\item{flag_case}{TBA}

\item{flag_sero}{TBA}

\item{year0}{First year in population/vaccination data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRVC input in list from previous run(s)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{dt}{Time increment in days to use in model (should be either 1.0 or 5.0 days)
'}
}
\description{
Run SEIRV model for one region when dataset being generated using multithreading
}
\details{
Accepts epidemiological + population parameters and model settings; runs SEIRV model
for one region over a specified time period for a number of particles/threads and outputs specific data for dataset
being generated in parallel.
}
