% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{packer_setup}
\alias{packer_setup}
\title{packer_setup}
\usage{
packer_setup(
  pars_fixed = list(),
  env_covar_values = list(),
  mode_time = 0,
  vars_extra_names = c("")
)
}
\arguments{
\item{pars_fixed}{Fixed parameters list created using pars_fixed_setup()}

\item{env_covar_values}{Environmental covariate values (TBC)}

\item{mode_time}{Type of time dependence of FOI_spillover and R0 to be used: \cr
If mode_time = 0, no time variation (constant values)\cr
If mode_time = 1, FOI/R0 vary annually without seasonality (number of values = number of years to consider) \cr
If mode_time = 2, FOI/R0 vary with monthly seasonality without inter - annual variation (number of values = 12) \cr
If mode_time = 3, FOI/R0 vary with daily seasonality without inter - annual variation (number of values = 365/time_inc) \cr
If mode_time = 4, FOI/R0 vary annually with monthly seasonality (number of values = 12*number of years to consider) \cr
If mode_time = 5, FOI/R0 vary annually with daily seasonality (number of values = (365/time_inc)*number of years to consider)}

\item{vars_extra_names}{Names of additional varied parameters}
}
\description{
Create packer for fitting
}
\details{
TBA
}
