% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_nolog.R
\name{param_prop_setup2}
\alias{param_prop_setup2}
\title{param_prop_setup2}
\usage{
param_prop_setup2(params = c(), chain_cov = 1, adapt = 0)
}
\arguments{
\item{params}{Previous parameter values used as input}

\item{chain_cov}{Covariance calculated from previous iterations in chain}

\item{adapt}{0/1 flag indicating which type of covariance to use for proposition value (TBA)
'}
}
\description{
Set up proposed new parameter values for next iteration in chain
}
\details{
Takes in current values of parameter set used for Markov Chain Monte Carlo fitting and proposes new values
from multivariate normal distribution where the existing values form the mean and the standard deviation is
based on the chain covariance or (if the flag "adapt" is set to 1) a flat value based on the number of parameters.
}
