% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{pars_var_setup}
\alias{pars_var_setup}
\title{pars_var_setup}
\usage{
pars_var_setup(
  n_env_vars = 5,
  vars_extra_names = c("p_rep_severe", "p_rep_death")
)
}
\arguments{
\item{n_env_vars}{TBA}

\item{vars_extra_names}{TBA}
}
\description{
TBA
}
\details{
TBA
}
