% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_tf.R
\name{precip_tf}
\alias{precip_tf}
\title{precip_tf}
\usage{
precip_tf(precip_values = c(0), a_ptf = 0)
}
\arguments{
\item{precip_values}{TBA}

\item{a_ptf}{TBA
'}
}
\description{
Function for precipitation transformation
}
\details{
TBA
}
