% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{prior_setup}
\alias{prior_setup}
\title{prior_setup}
\usage{
prior_setup(
  packer = NULL,
  env_covar_values = list(),
  pars_var = list(),
  FOI_R0_prior_data = list()
)
}
\arguments{
\item{packer}{Packer created using packer_setup()}

\item{env_covar_values}{Environmental covariate values (TBC)}

\item{pars_var}{Data frame of information on varied parameters, created using pars_var_setup()}

\item{FOI_R0_prior_data}{Data frame of prior data for FOI/R0}
}
\description{
Set up prior function for fitting
}
\details{
TBA
}
