% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_tf.R
\name{quad}
\alias{quad}
\title{quad}
\usage{
quad(temp_values = c(25), T0 = 12.71508, Tm = 38.04809469, c = -0.757869)
}
\arguments{
\item{temp_values}{TBA}

\item{T0}{TBA}

\item{Tm}{TBA}

\item{c}{TBA
'}
}
\description{
Quadratic function for temperature transformation
}
\details{
TBA
}
