% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{regions_breakdown}
\alias{regions_breakdown}
\title{regions_breakdown}
\usage{
regions_breakdown(region_labels = c())
}
\arguments{
\item{region_labels}{Vector of region labels}
}
\description{
Break down regions from datasets to get list of unique regions
}
\details{
Takes in vector of region labels, potentially including labels which contain more than one region (e.g.
  labels for countrywide data with all adm1 level regions as one label separated by comma) and produces an alphabetically
  ordered list of region labels, breaking down comma-separated groups.

  For example, supplying a vector where all labels are "BRA.1_1,BRA.2_1,BRA.3_1" will return a vector of length 3
  - c("BRA.1_1","BRA.2_1","BRA.3_1").
}
