% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero.R
\name{sero_calculate2_alt}
\alias{sero_calculate2_alt}
\title{sero_calculate2_alt}
\usage{
sero_calculate2_alt(
  sero_data = list(),
  model_data = list(),
  n_region = 1,
  n_p = 1
)
}
\arguments{
\item{sero_data}{Data frame containing years, minimum and maximum ages, and values of vc_factor (proportion of
people for whom vaccination status unknown)}

\item{model_data}{Annual cumulative SEIRV output of Model_Run}

\item{n_region}{Region number to select from model_data}

\item{n_p}{Particle to select from model_data
'}
}
\description{
Calculate number of "samples" and number of "positives" from modelled data for specified age range(s)
and year(s)
}
\details{
Takes in information on minimum and maximum ages of desired range(s), year(s) for which to calculate
number of "samples" (people eligible for testing) and "positives" (people who would test positive), plus vc_factor
(proportion of people for whom vaccination status unknown)
}
