% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero.R
\name{sero_data_compare}
\alias{sero_data_compare}
\title{sero_data_compare}
\usage{
sero_data_compare(model_sero_values = c(), obs_sero_data = list())
}
\arguments{
\item{model_sero_values}{Seroprevalence values from dataset generation function (no. positives/no. samples)}

\item{obs_sero_data}{Seroprevalence data for comparison, by year and age group, in format
  no. samples/no. positives
'}
}
\description{
Take seroprevalence results from dataset generation functions, compare comparison with
observed/template seroprevalence data and calculate likelihood
}
\details{
[TBA]
}
