% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc2.R
\name{single_posterior_calc2}
\alias{single_posterior_calc2}
\title{single_posterior_calc2}
\usage{
single_posterior_calc2(
  log_params_prop = c(),
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  consts = list()
)
}
\arguments{
\item{log_params_prop}{TBA}

\item{input_data}{TBA}

\item{obs_sero_data}{TBA}

\item{obs_case_data}{TBA}

\item{consts}{TBA}
}
\description{
TBA
}
\details{
TBA
}
