% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_tf.R
\name{temp_tf}
\alias{temp_tf}
\title{temp_tf}
\usage{
temp_tf(
  temp_values = c(25),
  a_T0 = 2.248952,
  a_Tm = 40.13383,
  a_c = 0.000271964,
  mu_T0 = 12.71508,
  mu_Tm = 38.04809469,
  mu_c = -0.757869,
  PDR_T0 = 17.33263,
  PDR_Tm = 42.19592,
  PDR_c = 0.000135891
)
}
\arguments{
\item{temp_values}{TBA}

\item{a_T0}{TBA}

\item{a_Tm}{TBA}

\item{a_c}{TBA}

\item{mu_T0}{TBA}

\item{mu_Tm}{TBA}

\item{mu_c}{TBA}

\item{PDR_T0}{TBA}

\item{PDR_Tm}{TBA}

\item{PDR_c}{TBA
'}
}
\description{
Function for temperature transformation
}
\details{
TBA
}
