% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{template_region_xref}
\alias{template_region_xref}
\title{template_region_xref}
\usage{
template_region_xref(template = list(), regions = c())
}
\arguments{
\item{template}{[TBA]}

\item{regions}{[TBA]}
}
\description{
[TBA]
}
\details{
[TBA]
}
