% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_tf.R
\name{epi_param_calc2}
\alias{epi_param_calc2}
\title{epi_param_calc2}
\usage{
epi_param_calc2(pars_fixed = list(), env_covar_values = list(), p = list())
}
\arguments{
\item{pars_fixed}{TBA}

\item{env_covar_values}{TBA}

\item{p}{TBA
'}
}
\description{
TBA
}
\details{
TBA
}
