% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{fit_data_setup}
\alias{fit_data_setup}
\title{fit_data_setup}
\usage{
fit_data_setup(
  sero_template = list(),
  case_template = list(),
  year0 = 1940,
  time_inc = 5,
  region_index_sero = list(),
  region_index_case = list()
)
}
\arguments{
\item{sero_template}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{case_template}{Annual reported case/death data for comparison, by region and year, in format no.
cases/no. deaths}

\item{year0}{First year in population/vaccination data}

\item{time_inc}{Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)}

\item{region_index_sero}{TBA}

\item{region_index_case}{TBA}
}
\description{
Create fitting dataset from serological and case datasets
}
\details{
TBA
}
