% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{fit_data_setup}
\alias{fit_data_setup}
\title{fit_data_setup}
\usage{
fit_data_setup(
  sero_template = list(),
  case_template = list(),
  year0 = 1940,
  time_inc = 5,
  region_index_sero = list(),
  region_index_case = list()
)
}
\arguments{
\item{sero_template}{TBA}

\item{case_template}{TBA}

\item{year0}{TBA}

\item{time_inc}{TBA}

\item{region_index_sero}{TBA}

\item{region_index_case}{TBA}
}
\description{
TBA
}
\details{
TBA
}
