% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_generate.R
\name{Generate_Dataset}
\alias{Generate_Dataset}
\title{Generate_Dataset}
\usage{
Generate_Dataset(
  input_data = list(),
  FOI_values = c(),
  R0_values = c(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  vaccine_efficacy = 1,
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe = 1,
  p_rep_death = 1,
  mode_start = 1,
  start_SEIRV = NULL,
  dt = 1,
  n_reps = 1,
  deterministic = FALSE,
  mode_parallel = "none",
  cluster = NULL
)
}
\arguments{
\item{input_data}{List of population and vaccination data for multiple regions}

\item{FOI_values}{Values for each region of the force of infection due to spillover from sylvatic reservoir}

\item{R0_values}{Values for each region of the basic reproduction number for human-human transmission}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no. cases/no.
deaths}

\item{vaccine_efficacy}{Fractional vaccine efficacy}

\item{p_severe_inf}{Probability of an infection being severe}

\item{p_death_severe_inf}{Probability of a severe infection resulting in death}

\item{p_rep_severe}{Probability of reporting of a severe but non-fatal infection}

\item{p_rep_death}{Probability of reporting of a fatal infection}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input (list of datasets, one per region)}

\item{dt}{Time increment in days to use in model (should be either 1.0 or 5.0 days)}

\item{n_reps}{number of stochastic repetitions}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE}

\item{mode_parallel}{Set mode for parallelization, if any:
If mode_parallel="none", no parallelization
If mode_parallel="pars_multi", all regions run in parallel for same time period with same output type
If mode_parallel="clusterMap", all regions run in parallel with different time periods and output types}

\item{cluster}{Cluster of threads to use if mode_parallel="clusterMap"
'}
}
\description{
Generate serological and/or annual case/death data
}
\details{
This function is used to generate serological and/or annual case/death based on observed or dummy data sets;
it is normally used by the single_like_calc() and data_match_single() functions.
}
