% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc2.R
\name{Generate_Dataset2}
\alias{Generate_Dataset2}
\title{Generate_Dataset2}
\usage{
Generate_Dataset2(
  input_data = list(),
  FOI_values = c(),
  R0_values = c(),
  sero_template = NULL,
  case_template = NULL,
  vaccine_efficacy = 1,
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe_af = 1,
  p_rep_death_af = 1,
  p_rep_severe_sa = 1,
  p_rep_death_sa = 1,
  mode_start = 1,
  start_SEIRV = NULL,
  dt = 1,
  n_reps = 1,
  deterministic = FALSE,
  mode_parallel = "none",
  cluster = NULL,
  output_frame = FALSE
)
}
\arguments{
\item{input_data}{TBA}

\item{FOI_values}{TBA}

\item{R0_values}{TBA}

\item{sero_template}{TBA}

\item{case_template}{TBA}

\item{vaccine_efficacy}{TBA}

\item{p_severe_inf}{TBA}

\item{p_death_severe_inf}{TBA}

\item{p_rep_severe_af}{TBA}

\item{p_rep_death_af}{TBA}

\item{p_rep_severe_sa}{TBA}

\item{p_rep_death_sa}{TBA}

\item{mode_start}{TBA}

\item{start_SEIRV}{TBA}

\item{dt}{TBA}

\item{n_reps}{TBA}

\item{deterministic}{TBA}

\item{mode_parallel}{TBA}

\item{cluster}{TBA}

\item{output_frame}{TBA
'}
}
\description{
TBA
}
\details{
TBA
}
