% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_gen_VIMC.R
\name{Generate_VIMC_Burden_Dataset}
\alias{Generate_VIMC_Burden_Dataset}
\title{Generate_VIMC_Burden_Dataset}
\usage{
Generate_VIMC_Burden_Dataset(
  FOI_values = c(),
  R0_values = c(),
  input_data = list(),
  template = NULL,
  vaccine_efficacy = 1,
  time_inc = 1,
  mode_start = 1,
  start_SEIRV = NULL,
  mode_time = 0,
  n_reps = 1,
  deterministic = FALSE,
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  YLD_per_case = 0.006486,
  mode_div = "default",
  n_nodes = 1,
  seed = NULL,
  xref = NULL
)
}
\arguments{
\item{FOI_values}{Array of values of force of infection due to spillover from sylvatic reservoir by region + time point}

\item{R0_values}{Array of values of basic reproduction number for human-human transmission by region and time point}

\item{input_data}{List of population and vaccination data for multiple regions}

\item{template}{Burden data in VIMC format, with regions, years, minimum and maximum age, and life expectancy
for each line}

\item{vaccine_efficacy}{Fractional vaccine efficacy}

\item{time_inc}{Time increment in days to use in model (should be either 1.0, 2.5 or 5.0 days)}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination \cr
If mode_start = 0, only vaccinated individuals \cr
If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity (stratified by age) \cr}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input (list of datasets, one per region)}

\item{mode_time}{Type of time dependence of FOI_spillover and R0 to be used: \cr
If mode_time = 0, no time variation (constant values)\cr
If mode_time = 1, FOI/R0 vary annually without seasonality (number of values = number of years to consider) \cr
If mode_time = 2, FOI/R0 vary with monthly seasonality without inter-annual variation (number of values = 12) \cr
If mode_time = 3, FOI/R0 vary with daily seasonality without inter-annual variation (number of values = 365/dt) \cr
If mode_time = 4, FOI/R0 vary annually with monthly seasonality (number of values = 12*number of years to consider) \cr
If mode_time = 5, FOI/R0 vary annually with daily seasonality (number of values = (365/dt)*number of years to consider)}

\item{n_reps}{number of stochastic repetitions}

\item{deterministic}{TRUE/FALSE - set model to run in deterministic mode if TRUE}

\item{p_severe_inf}{Probability of an infection being severe}

\item{p_death_severe_inf}{Probability of a severe infection resulting in death}

\item{YLD_per_case}{TBA}

\item{mode_div}{Mode of division of regions while running model. Parallel mode will run all years for every region \cr
If mode_div = "parallel", run all regions simultaneously in a single use of the Model_Run function \cr
If mode_div = "serial", run all regions in series \cr
If mode_div = "cluster", set model to run regions in parallel using cluster of nodes (see n_nodes parameter)}

\item{n_nodes}{Number of nodes to use in parallelization (if mode_div = cluster)}

\item{seed}{Optional random seed value; set to NULL to omit. \cr
If mode_div = serial, same seed is used for all regions}

\item{xref}{Region cross-referencing data generated by template_region_xref() function\cr
  If set to NULL, calculated within function. Pre-calculate if running multiple times with same template,
  to reduce computation time.
'}
}
\description{
Generate annual burden data in format used by VIMC
}
\details{
This function is used to generate annual burden data in the format used by the Vaccine Impact
  Modelling Consortium (VIMC) [TBA]
}
