% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{get_region_grouping}
\alias{get_region_grouping}
\title{get_region_grouping}
\usage{
get_region_grouping(
  regions = c(),
  sero_template = list(),
  case_template = list(),
  xref_sero = list(),
  xref_case = list(),
  flag_simple = TRUE
)
}
\arguments{
\item{regions}{TBA}

\item{sero_template}{TBA}

\item{case_template}{TBA}

\item{xref_sero}{TBA}

\item{xref_case}{TBA}

\item{flag_simple}{TBA
'}
}
\description{
Group regions from serology and case data templates
}
\details{
TBA
}
