% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{get_region_grouping}
\alias{get_region_grouping}
\title{get_region_grouping}
\usage{
get_region_grouping(
  regions = c(),
  template = list(sero = NULL, case = NULL, xref_sero = NULL, xref_case = NULL),
  mode_grouping = 1
)
}
\arguments{
\item{regions}{TBA}

\item{template}{TBA}

\item{mode_grouping}{TBA
'}
}
\description{
Group regions from serology and case data template
}
\details{
TBA
}
