% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{m_prelim_fit}
\alias{m_prelim_fit}
\title{m_prelim_fit}
\usage{
m_prelim_fit(
  fit_data = list(),
  packer = NULL,
  prior = NULL,
  FOI_R0_prior_data = list(),
  pars_var = list(),
  env_covar_values = list(),
  n_values = 10,
  n_particles = 1,
  n_threads = 1
)
}
\arguments{
\item{fit_data}{TBA}

\item{packer}{TBA}

\item{prior}{TBA}

\item{FOI_R0_prior_data}{TBA}

\item{pars_var}{TBA}

\item{env_covar_values}{TBA}

\item{n_values}{TBA}

\item{n_particles}{TBA}

\item{n_threads}{TBA}
}
\description{
TBA
}
\details{
TBA
}
