% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{m_sample}
\alias{m_sample}
\title{m_sample}
\usage{
m_sample(
  fit_data = list(),
  packer = NULL,
  prior = NULL,
  FOI_R0_prior_data = list(),
  pars_var = list(),
  initial = list(),
  vcv = list(),
  rerun_every = 50,
  n_chains = 1,
  n_iterations = 10,
  deterministic = FALSE,
  n_particles = 1,
  n_threads = 1,
  parallel = FALSE,
  output_file = NULL,
  seed = NULL
)
}
\arguments{
\item{fit_data}{Fitting data created using fit_data_setup()}

\item{packer}{Packer created using packer_setup()}

\item{prior}{Prior function created using prior_setup()}

\item{FOI_R0_prior_data}{Data frame of prior data for FOI/R0}

\item{pars_var}{Data frame of information on varied parameters, created using pars_var_setup()}

\item{initial}{TBA}

\item{vcv}{Variance-covariance matrix}

\item{rerun_every}{TBA}

\item{n_chains}{Number of chains}

\item{n_iterations}{Number of iterations for which to run each chain}

\item{deterministic}{TBA}

\item{n_particles}{Number of particles}

\item{n_threads}{Number of threads}

\item{parallel}{TBA}

\item{output_file}{Name of file location to save results (set to NULL if unused)}

\item{seed}{Random seed (set to NULL if unused)}
}
\description{
TBA
}
\details{
TBA
}
