% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc2.R
\name{MCMC2}
\alias{MCMC2}
\title{MCMC2}
\usage{
MCMC2(
  log_params_ini = c(),
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  filename_prefix = "Chain",
  Niter = 1,
  type = NULL,
  log_params_min = c(),
  log_params_max = c(),
  mode_start = 0,
  prior_type = "zero",
  norm_prior_sd = c(30, 10, 30),
  dt = 1,
  n_reps = 1,
  enviro_data = NULL,
  R0_fixed_values = NULL,
  vaccine_efficacy = NULL,
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe_af = NULL,
  p_rep_death_af = NULL,
  p_rep_severe_sa = NULL,
  p_rep_death_sa = NULL,
  m_FOI_Brazil = 1,
  deterministic = FALSE,
  mode_parallel = "none",
  cluster = NULL
)
}
\arguments{
\item{log_params_ini}{TBA}

\item{input_data}{TBA}

\item{obs_sero_data}{TBA}

\item{obs_case_data}{TBA}

\item{filename_prefix}{TBA}

\item{Niter}{TBA}

\item{type}{TBA}

\item{log_params_min}{TBA}

\item{log_params_max}{TBA}

\item{mode_start}{TBA}

\item{prior_type}{TBA}

\item{norm_prior_sd}{TBA}

\item{dt}{TBA}

\item{n_reps}{TBA}

\item{enviro_data}{TBA}

\item{R0_fixed_values}{TBA}

\item{vaccine_efficacy}{TBA}

\item{p_severe_inf}{TBA}

\item{p_death_severe_inf}{TBA}

\item{p_rep_severe_af}{TBA}

\item{p_rep_death_af}{TBA}

\item{p_rep_severe_sa}{TBA}

\item{p_rep_death_sa}{TBA}

\item{m_FOI_Brazil}{TBA}

\item{deterministic}{TBA}

\item{mode_parallel}{TBA}

\item{cluster}{TBA
'}
}
\description{
TBA
}
\details{
TBA
}
