% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_checks}
\alias{mcmc_checks}
\title{mcmc_checks}
\usage{
mcmc_checks(params_data = list(), covar_names = c(), check_initial = FALSE)
}
\arguments{
\item{params_data}{TBA}

\item{covar_names}{TBA}

\item{check_initial}{TBA}
}
\description{
Perform checks on parameter data
}
\details{
This function, which is called by MCMC() and mcmc_prelim_fit(), performs a number of checks on parameter data
to ensure proper functionality. [TBA]
}
