% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_pa.R
\name{mcmc_checks2}
\alias{mcmc_checks2}
\title{mcmc_checks2}
\usage{
mcmc_checks2(params_data = list(), covar_names = c())
}
\arguments{
\item{params_data}{TBA}

\item{covar_names}{TBA}
}
\description{
Perform checks on parameter data
}
\details{
This function, which is called by MCMC(), performs a number of checks on parameter data to
ensure proper functionality. [TBA]
}
