% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_nolog.R
\name{mcmc_checks2}
\alias{mcmc_checks2}
\title{mcmc_checks2}
\usage{
mcmc_checks2(
  params_ini = c(),
  n_regions = 1,
  prior_settings = list(type = "zero"),
  enviro_data_const = list(),
  enviro_data_var = list(),
  add_values = list(vaccine_efficacy = 1, p_rep_severe = 1, p_rep_death = 1, m_FOI_Brazil
    = 1),
  extra_estimated_params = c()
)
}
\arguments{
\item{params_ini}{Initial values of parameters to be estimated (see documentation for MCMC2() function for more details)}

\item{n_regions}{Number of regions}

\item{prior_settings}{List containing settings for priors; see documentation for MCMC() function for more details)}

\item{enviro_data_const}{Data frame of values of constant environmental covariates (columns) by region (rows)}

\item{enviro_data_var}{List containing values of time-varying environmental covariates (TBA)}

\item{add_values}{List of parameters in addition to those governing FOI/R0, either giving a fixed value or giving NA to
 indicate that they are part of the parameter set to be estimated \cr
vaccine_efficacy Vaccine efficacy (proportion of reported vaccinations causing immunity) (must be present) \cr
p_rep_severe Probability of observation of severe infection \cr
p_rep_death Probability of observation of death \cr
m_FOI_Brazil Multiplier of spillover FOI for Brazil regions (only relevant if regions in Brazil to be considered) \cr}

\item{extra_estimated_params}{Vector of names of parameters to be estimated in addition to those governing FOI and R0;
see add_values}
}
\description{
Perform checks on MCMC inputs
}
\details{
This function, which is called by MCMC2(), performs a number of checks on data to be used in fitting to
ensure proper functionality. It verifies that the number of parameters being estimated is consistent with other
settings and that certain values are not outwith sensible boundaries (e.g. probabilities must be between 0 and 1).
}
