% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_FOI_R0_setup}
\alias{mcmc_FOI_R0_setup}
\title{mcmc_FOI_R0_setup}
\usage{
mcmc_FOI_R0_setup(regions = "", log_params_prop = c(), enviro_data = list())
}
\arguments{
\item{regions}{Vector of region names}

\item{log_params_prop}{Proposed values of parameters (natural logarithm of actual parameters)}

\item{enviro_data}{Environmental data frame, containing only relevant environmental covariate values
'}
}
\description{
Set up FOI and R0 values and calculate some prior probability values for MCMC calculation
}
\details{
Takes in parameter values used for Markov Chain Monte Carlo fitting, calculates spillover force of
infection and (optionally) reproduction number values either directly or from environmental covariates. Also
calculates related components of prior probability.
}
