% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{packer_setup}
\alias{packer_setup}
\title{packer_setup}
\usage{
packer_setup(pars_fixed = list(), env_covar_values = list(), mode_time = 0)
}
\arguments{
\item{pars_fixed}{TBA}

\item{env_covar_values}{TBA}

\item{mode_time}{TBA}
}
\description{
TBA
}
\details{
TBA
}
