% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parameter_setup}
\alias{parameter_setup}
\title{Parameter setup}
\usage{
parameter_setup(
  FOI_spillover = c(),
  R0 = c(),
  vacc_data = list(),
  pop_data = list(),
  year0 = 1940,
  years_data = c(),
  mode_start = 0,
  vaccine_efficacy = 1,
  start_SEIRV = list(),
  time_inc = 1,
  mode_time = 0
)
}
\arguments{
\item{FOI_spillover}{Force of infection due to spillover from sylvatic reservoir (TBA)}

\item{R0}{Reproduction number for urban spread of infection (TBA)}

\item{vacc_data}{Projected vaccination-based immunity (assuming vaccine_efficacy = 1) by age group and year}

\item{pop_data}{Population by age group and year}

\item{year0}{First year in population/vaccination data}

\item{years_data}{Incremental vector of years denoting years for which to save data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination \cr
If mode_start = 0, only vaccinated individuals \cr
If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity (uniform by age,
R0 based only) \cr
If mode_start = 2, use SEIRV input in list from previous run(s) \cr
If mode_start = 3, shift some non-vaccinated individuals into recovered to give herd immunity (stratified by age)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{time_inc}{Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)}

\item{mode_time}{Type of time dependence of FOI_spillover and R0 to be used: \cr
 If mode_time=0, no time variation (constant values)\cr
 If mode_time=1, FOI/R0 vary annually without seasonality (number of values = number of years to consider) \cr
 If mode_time=2, FOI/R0 vary with monthly seasonality without inter-annual variation (number of values = 12) \cr
 If mode_time=3, FOI/R0 vary with daily seasonality without inter-annual variation (number of values = 365/dt) \cr
 If mode_time=4, FOI/R0 vary annually with monthly seasonality (number of values = 12*number of years to consider) \cr
 If mode_time=5, FOI/R0 vary annually with daily seasonality (number of values = (365/dt)*number of years to consider)
'}
}
\description{
Set up parameters to input into model
}
\details{
Takes in multiple inputs, outputs list for use by odin SEIRV model.
}
