% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parameter_setup}
\alias{parameter_setup}
\title{Parameter setup}
\usage{
parameter_setup(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  year0 = 1940,
  years_data = c(1941:1942),
  mode_start = 0,
  vaccine_efficacy = 1,
  start_SEIRV = list(),
  dt = 1
)
}
\arguments{
\item{FOI_spillover}{Force of infection due to spillover from sylvatic reservoir}

\item{R0}{Reproduction number for urban spread of infection}

\item{vacc_data}{Projected vaccination-based immunity (assuming vaccine_efficacy=1) by age group and year}

\item{pop_data}{Population by age group and year}

\item{year0}{First year in population/vaccination data}

\item{years_data}{Incremental vector of years denoting years for which to save data}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start=0, only vaccinated individuals
If mode_start=1, shift some non-vaccinated individuals into recovered to give herd immunity
If mode_start=2, use SEIRV input in list from previous run(s)}

\item{vaccine_efficacy}{Proportional vaccine efficacy}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input}

\item{dt}{Time increment in days to use in model (should be either 1.0 or 5.0 days)
'}
}
\description{
Set up parameters to input into model
}
\details{
Takes in multiple inputs, outputs list for use by odin SEIRV model.
}
