% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{pars_fixed_setup}
\alias{pars_fixed_setup}
\title{pars_fixed_setup}
\usage{
pars_fixed_setup(
  sero_template = list(),
  case_template = list(),
  vacc_data = list(),
  pop_data = list(),
  years_data = c(),
  year0 = 1940,
  time_inc = 1,
  mode_start = 0,
  start_SEIRV = NULL,
  fixed_extra = list(),
  ref_BRA = NULL,
  i_ttf = NULL,
  i_ptf = NULL
)
}
\arguments{
\item{sero_template}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{case_template}{Annual reported case/death data for comparison, by region and year, in format no.
cases/no. deaths}

\item{vacc_data}{Projected vaccination-based immunity (assuming vaccine_efficacy = 1) by region. age group and year}

\item{pop_data}{Population by region, age group and year}

\item{years_data}{Vector of years denoting years for which data needed (TBC?)}

\item{year0}{First year in population/vaccination data}

\item{time_inc}{Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)}

\item{mode_start}{Flag indicating how to set initial population immunity level in addition to vaccination \cr
If mode_start = 0, only vaccinated individuals \cr
If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity (stratified by age) \cr
If mode_start = 2, use SEIRV input in list from previous run(s) (TBD) \cr}

\item{start_SEIRV}{SEIRV data from end of a previous run to use as input (if mode_start = 2)}

\item{fixed_extra}{List containing additional fixed parameters (from list available\cr
as variable parameters)}

\item{ref_BRA}{List of region numbers for which Brazil FOI multiplier to be applied}

\item{i_ttf}{Number of environmental covariate for which temperature transformation to be applied}

\item{i_ptf}{Number of environmental covariate for which precipitation transformation to be applied}
}
\description{
Set up fixed parameters for fitting
}
\details{
TBA
}
