% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{pars_fixed_setup}
\alias{pars_fixed_setup}
\title{pars_fixed_setup}
\usage{
pars_fixed_setup(
  sero_template = list(),
  case_template = list(),
  vacc_data = list(),
  pop_data = list(),
  years_data = c(),
  year0 = 1940,
  time_inc = 1,
  mode_start = 0,
  start_SEIRV = NULL,
  fixed_extra = list(),
  ref_BRA = NULL
)
}
\arguments{
\item{sero_template}{TBA}

\item{case_template}{TBA}

\item{vacc_data}{TBA}

\item{pop_data}{TBA}

\item{years_data}{TBA}

\item{year0}{TBA}

\item{time_inc}{TBA}

\item{mode_start}{TBA}

\item{start_SEIRV}{TBA}

\item{fixed_extra}{TBA}

\item{ref_BRA}{TBA}
}
\description{
TBA
}
\details{
TBA
}
