% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_monty.R
\name{prior_setup}
\alias{prior_setup}
\title{prior_setup}
\usage{
prior_setup(
  packer = NULL,
  env_covar_values = list(),
  pars_var = list(),
  FOI_R0_prior_data = list()
)
}
\arguments{
\item{packer}{TBA}

\item{env_covar_values}{TBA}

\item{pars_var}{TBA}

\item{FOI_R0_prior_data}{TBA}
}
\description{
TBA
}
\details{
TBA
}
